Update authentication flow to use refresh tokens
These will make the app more secure by ensuring the access tokens only
have a short life, and will enable a way to revoke all refresh tokens if a
comprimised refresh token is sent. This is currently a basic flow and
can be iterated on later. I am keeping the authentication in this app
for now but can also be moved to a specific authentication server later.

Update SALT_ROUNDS for hashing to be an optional env var that can be
set.