Create initial tests for api routes
Have a seperate Postgres container exposing a different port so
the tests don't affect the dev environment. Create some basic make
commands to simplify running these two envs. This is a first iteration
and can be improved on later.

Improve the logic for storing the refresh token expiry date in the db.
Previously it had the potential to be a second out as I was creating two
seperate dates and not using the one set in the refresh token.